Remove `fsnotify` dependency.
This marks a transition back to the System.INotify package. I didn't realize
that fsnotify was last touched back in 2015 and it has one single breaking
change (that's broken a lot of other projects depending on it) which is that
it uses an older FilePath library that's now deprecated.

One could fix it and submit a pull request, but I'd rather focus on getting this
up and running for Linux + OSX.