Implement initial command line flags
This provides two basic modes and some basic command line args: a 'watch' mode
for watching a single directory with a single task or a 'triggers' mode which
takes an implicit or explicit config file and sets up triggers for it. At the
moment the latter mode is not yet implemented. This sets up the codebase for
additional flags (verbosity, recursive searches, et. al.)

This resolves #7.